+++ src/index.ts
@@ -18,3 +18,4 @@ export { getErrorResponse, isRequestError, type RequestError } from "./RequestEr
 export * from "./UtilTypes";
 export { default as retry } from "./util/retry";
 export type { RetryOptions } from "./util/retry/interfaces";
+export { schema } from "./Validation";